[INTERNAL] move to scoring sub-module instead of data_structure
