+++ include/seqan3/alignment/scoring/scoring_scheme_base.hpp
@@ -59,7 +59,7 @@ namespace seqan3
 
 /*!\brief A strong type of underlying type `score_type` that represents the score of two matching characters.
  * \tparam score_type The underlying type.
- * \ingroup data_structure
+ * \ingroup scoring
  * \see scoring_scheme_base::set_simple_scheme
  */
 template <arithmetic_concept score_type>
@@ -82,7 +82,7 @@ match_score(score_type &&) -> match_score<score_type>;
 
 /*!\brief A strong type of underlying type `score_type` that represents the score two different characters.
  * \tparam score_type The underlying type.
- * \ingroup data_structure
+ * \ingroup scoring
  * \see scoring_scheme_base::set_simple_scheme
  */
 template <arithmetic_concept score_type>
@@ -107,7 +107,7 @@ mismatch_score(score_type &&) -> mismatch_score<score_type>;
  * \tparam derived_t  The derived type.
  * \tparam alphabet_t Type of the largest target alphabet.
  * \tparam score_type Type of the score values in the internal matrix.
- * \ingroup data_structure
+ * \ingroup scoring
  *
  * \details
  *
